Move SampleReads to correct location in multi-module build.
